refactor: merge admin deployment search back into the main layers
Fold DeploymentAdminRepository / DeploymentAdminService /
DeploymentAdminProcessors back into DeploymentRepository /
DeploymentService / DeploymentProcessors. admin_search_deployments now
lives alongside the scoped search methods, distinguished only by its
admin_ prefix and the absence of a scope filter — no dedicated admin
classes. Drops the now-unused DEPLOYMENT_ADMIN_REPOSITORY metric layer.

Also fix two rebase-induced CI breakages:
- main renamed execute_batch_querier's scope= kwarg to scopes=; update
  the user/project scoped search calls to scopes=[scope].
- repoint the new sokovan deploying_rolling_back handler (and its test)
  at the common DeploymentLifecycleSubStep location, since this branch
  moved the enum off manager.data.deployment.types.
