+++ src/ai/backend/manager/repositories/deployment/__init__.py
@@ -3,11 +3,9 @@
 from ai.backend.manager.models.endpoint.conditions import DeploymentConditions
 from ai.backend.manager.models.routing.conditions import RouteConditions
 
-from .admin_repository import DeploymentAdminRepository
 from .repository import DeploymentRepository
 
 __all__ = [
-    "DeploymentAdminRepository",
     "DeploymentRepository",
     "DeploymentConditions",
     "RouteConditions",
