+++ tests/unit/manager/sokovan/deployment/handlers/test_deploying_rolling_back_handler.py
@@ -9,11 +9,11 @@ import pytest
 from dateutil.tz import tzutc
 
 from ai.backend.common.data.endpoint.types import EndpointLifecycle, ScalingState
+from ai.backend.common.data.model_deployment.types import DeploymentLifecycleSubStep
 from ai.backend.common.identifier.deployment import DeploymentID
 from ai.backend.common.identifier.deployment_revision import DeploymentRevisionID
 from ai.backend.manager.data.deployment.types import (
     DeploymentInfo,
-    DeploymentLifecycleSubStep,
     DeploymentMetadata,
     DeploymentNetworkData,
     DeploymentOptions,
